+++ CHANGELOG.md
@@ -18,6 +18,8 @@ This beta release is built on top of [OpenTelemetry .NET](https://github.com/ope
 - Support configuring `OTEL_*` settings using `App.config` and `Web.config`.
 - Add support for Quartz traces instrumentation.
 - Add support for EntityFrameworkCore traces instrumentations.
+- Add plugin support for
+  `ResourceBuilder ConfigureResource(ResourceBuilder builder)`.
 
 ### Changed